# internal/api/authapi/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package authapi

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn4AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn3AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn13AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn8AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn20AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn1AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn11AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/api/auth/"

			if l := len("/api/auth/"); len(elem) >= l && elem[0:l] == "/api/auth/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "api-keys"

				if l := len("api-keys"); len(elem) >= l && elem[0:l] == "api-keys" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleListAPIKeysRequest([0]string{}, elemIsEscaped, w, r)
					case "POST":
						s.handleCreateAPIKeyRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET,POST",
							allowedHeaders: rn4AllowedHeaders,
							acceptPost:     "application/json",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "DELETE":
							s.handleRevokeAPIKeyRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "DELETE",
								allowedHeaders: nil,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			case 'm': // Prefix: "me"

				if l := len("me"); len(elem) >= l && elem[0:l] == "me" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleGetCurrentIdentityRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: nil,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			case 'p': // Prefix: "password-reset/"

				if l := len("password-reset/"); len(elem) >= l && elem[0:l] == "password-reset/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'c': // Prefix: "confirm"

					if l := len("confirm"); len(elem) >= l && elem[0:l] == "confirm" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleConfirmPasswordResetRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn3AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

				case 'r': // Prefix: "request"

					if l := len("request"); len(elem) >= l && elem[0:l] == "request" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleRequestPasswordResetRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn13AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			case 'r': // Prefix: "refresh"

				if l := len("refresh"); len(elem) >= l && elem[0:l] == "refresh" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "POST":
						s.handleRefreshAccessTokenRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "POST",
							allowedHeaders: nil,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			case 's': // Prefix: "s"

				if l := len("s"); len(elem) >= l && elem[0:l] == "s" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'e': // Prefix: "essions"

					if l := len("essions"); len(elem) >= l && elem[0:l] == "essions" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch r.Method {
						case "GET":
							s.handleListSessionsRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: nil,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						// Param: "id"
						// Leaf parameter, slashes are prohibited
						idx := strings.IndexByte(elem, '/')
						if idx >= 0 {
							break
						}
						args[0] = elem
						elem = ""

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "DELETE":
								s.handleRevokeSessionRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "DELETE",
									allowedHeaders: nil,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					}

				case 'i': // Prefix: "ign-"

					if l := len("ign-"); len(elem) >= l && elem[0:l] == "ign-" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'i': // Prefix: "in"

						if l := len("in"); len(elem) >= l && elem[0:l] == "in" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "POST":
								s.handlePasswordSignInRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "POST",
									allowedHeaders: rn8AllowedHeaders,
									acceptPost:     "application/json",
									acceptPatch:    "",
								})
							}

							return
						}

					case 'o': // Prefix: "out"

						if l := len("out"); len(elem) >= l && elem[0:l] == "out" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "POST":
								s.handleSignOutRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "POST",
									allowedHeaders: nil,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					case 'u': // Prefix: "up"

						if l := len("up"); len(elem) >= l && elem[0:l] == "up" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "POST":
								s.handleSignUpRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "POST",
									allowedHeaders: rn20AllowedHeaders,
									acceptPost:     "application/json",
									acceptPatch:    "",
								})
							}

							return
						}

					}

				}

			case 'v': // Prefix: "verification/"

				if l := len("verification/"); len(elem) >= l && elem[0:l] == "verification/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'c': // Prefix: "confirm"

					if l := len("confirm"); len(elem) >= l && elem[0:l] == "confirm" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleConfirmEmailVerificationRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn1AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

				case 'r': // Prefix: "request"

					if l := len("request"); len(elem) >= l && elem[0:l] == "request" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleRequestEmailVerificationRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn11AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/api/auth/"

			if l := len("/api/auth/"); len(elem) >= l && elem[0:l] == "/api/auth/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "api-keys"

				if l := len("api-keys"); len(elem) >= l && elem[0:l] == "api-keys" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = ListAPIKeysOperation
						r.summary = "List API keys"
						r.operationID = "listAPIKeys"
						r.operationGroup = ""
						r.pathPattern = "/api/auth/api-keys"
						r.args = args
						r.count = 0
						return r, true
					case "POST":
						r.name = CreateAPIKeyOperation
						r.summary = "Create an API key"
						r.operationID = "createAPIKey"
						r.operationGroup = ""
						r.pathPattern = "/api/auth/api-keys"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "DELETE":
							r.name = RevokeAPIKeyOperation
							r.summary = "Revoke an API key"
							r.operationID = "revokeAPIKey"
							r.operationGroup = ""
							r.pathPattern = "/api/auth/api-keys/{id}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}

				}

			case 'm': // Prefix: "me"

				if l := len("me"); len(elem) >= l && elem[0:l] == "me" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = GetCurrentIdentityOperation
						r.summary = "Get the current identity"
						r.operationID = "getCurrentIdentity"
						r.operationGroup = ""
						r.pathPattern = "/api/auth/me"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'p': // Prefix: "password-reset/"

				if l := len("password-reset/"); len(elem) >= l && elem[0:l] == "password-reset/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'c': // Prefix: "confirm"

					if l := len("confirm"); len(elem) >= l && elem[0:l] == "confirm" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = ConfirmPasswordResetOperation
							r.summary = "Confirm a password reset token"
							r.operationID = "confirmPasswordReset"
							r.operationGroup = ""
							r.pathPattern = "/api/auth/password-reset/confirm"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 'r': // Prefix: "request"

					if l := len("request"); len(elem) >= l && elem[0:l] == "request" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = RequestPasswordResetOperation
							r.summary = "Request a password reset message"
							r.operationID = "requestPasswordReset"
							r.operationGroup = ""
							r.pathPattern = "/api/auth/password-reset/request"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			case 'r': // Prefix: "refresh"

				if l := len("refresh"); len(elem) >= l && elem[0:l] == "refresh" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "POST":
						r.name = RefreshAccessTokenOperation
						r.summary = "Rotate the refresh session and issue an access token"
						r.operationID = "refreshAccessToken"
						r.operationGroup = ""
						r.pathPattern = "/api/auth/refresh"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 's': // Prefix: "s"

				if l := len("s"); len(elem) >= l && elem[0:l] == "s" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'e': // Prefix: "essions"

					if l := len("essions"); len(elem) >= l && elem[0:l] == "essions" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch method {
						case "GET":
							r.name = ListSessionsOperation
							r.summary = "List active refresh sessions"
							r.operationID = "listSessions"
							r.operationGroup = ""
							r.pathPattern = "/api/auth/sessions"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						// Param: "id"
						// Leaf parameter, slashes are prohibited
						idx := strings.IndexByte(elem, '/')
						if idx >= 0 {
							break
						}
						args[0] = elem
						elem = ""

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "DELETE":
								r.name = RevokeSessionOperation
								r.summary = "Revoke a refresh session"
								r.operationID = "revokeSession"
								r.operationGroup = ""
								r.pathPattern = "/api/auth/sessions/{id}"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}

					}

				case 'i': // Prefix: "ign-"

					if l := len("ign-"); len(elem) >= l && elem[0:l] == "ign-" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'i': // Prefix: "in"

						if l := len("in"); len(elem) >= l && elem[0:l] == "in" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "POST":
								r.name = PasswordSignInOperation
								r.summary = "Sign in with email and password"
								r.operationID = "passwordSignIn"
								r.operationGroup = ""
								r.pathPattern = "/api/auth/sign-in"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					case 'o': // Prefix: "out"

						if l := len("out"); len(elem) >= l && elem[0:l] == "out" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "POST":
								r.name = SignOutOperation
								r.summary = "Sign out the current refresh session"
								r.operationID = "signOut"
								r.operationGroup = ""
								r.pathPattern = "/api/auth/sign-out"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					case 'u': // Prefix: "up"

						if l := len("up"); len(elem) >= l && elem[0:l] == "up" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "POST":
								r.name = SignUpOperation
								r.summary = "Create an account"
								r.operationID = "signUp"
								r.operationGroup = ""
								r.pathPattern = "/api/auth/sign-up"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					}

				}

			case 'v': // Prefix: "verification/"

				if l := len("verification/"); len(elem) >= l && elem[0:l] == "verification/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'c': // Prefix: "confirm"

					if l := len("confirm"); len(elem) >= l && elem[0:l] == "confirm" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = ConfirmEmailVerificationOperation
							r.summary = "Confirm an email verification token"
							r.operationID = "confirmEmailVerification"
							r.operationGroup = ""
							r.pathPattern = "/api/auth/verification/confirm"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 'r': // Prefix: "request"

					if l := len("request"); len(elem) >= l && elem[0:l] == "request" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = RequestEmailVerificationOperation
							r.summary = "Request an email verification message"
							r.operationID = "requestEmailVerification"
							r.operationGroup = ""
							r.pathPattern = "/api/auth/verification/request"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			}

		}
	}
	return r, false
}